list all movies
I finished the implementation for listing all the movies with a dict with some predefined values that I created in app.py. But there's nothing in the code yet that adds anything else to the dict. So that could be potentially be added in create new movie. I can also make the list all movies page look better too if y'all want.